Persist haikunated URLs in the DB
After a project with multiple services gets deployed, we'll get a list
of URLs from Dockup lib. Save these URLs into the service_urls JSON
field in Deployment model.  If the status argument received by run/4 is
"started", we have to save the payload (which will be the URLs of all
the services in our project) into our DB along with the deployment.