Accept options hash for DeployJob.perform
The project and deploy_job arguments are present for injecting
dependencies when testing, so they can be combined into an options
keyword list argument.

The callback is also used in a similar manner, but moving it into the
body of the function makes it unavailable in the rescue clause, so
keeping it as an argument. The alternative is to use a try-rescue block
inside the function after the callback binding has been defined in the
function, but that seems less elegant.